refactor: migrate field-label
