+++ packages/components/field-label/src/field-label.story.js
@@ -3,9 +3,9 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs/react';
 import FlatButton from '@commercetools-uikit/flat-button';
-import * as icons from '../icons';
-import Section from '../../../.storybook/decorators/section';
-import Readme from './README.md';
+import * as icons from '@commercetools-uikit/icons';
+import Section from '../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import FieldLabel from './field-label';
 
 storiesOf('Components|FieldLabel', module)
