+++ packages/components/field-label/src/field-label.visualroute.js
@@ -6,7 +6,7 @@ import {
   FlatButton,
   customProperties,
 } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../test/percy';
+import { Suite, Spec } from '../../../../test/percy';
 
 export const routePath = '/field-label';
