+++ packages/components/field-label/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './field-label';
+
+export { default as version } from './version';
