--- src/components/field-label/index.js
@@ -1,1 +1,0 @@
-export { default } from './field-label';
